Backend/requirements.txt
Flask==2.3.2
pandas
kafka-python==2.0.2
flask_cors